+++ core/apps.py
@@ -1,4 +1,6 @@
-from handlers import BoinkApp, TrackerApp, DefApp
+from handlers.boink_app import BoinkApp
+from handlers.tracker_app import TrackerApp
+from handlers.def_app import DefApp
 
 
 class Applications:
